добавлена сборка css по 2 уроку
