--- Students/Nina_Dokukina/Project/0/src/index.js
@@ -1,2 +1,0 @@
-import app from './js/main.js'
-app();
