--- Students/Nina_Dokukina/Project/0/src/js/main.js
@@ -1,4 +1,0 @@
-import cart from './cartComp.js'
-import catalog from './catalogComp.js'
-
- export default app
