+++ Students/Nina_Dokukina/Project/1/public/index.html
@@ -1,0 +1,49 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  
+    <title>Document</title>
+</head>
+
+<body>
+    <header>
+        <div class="logo">E-shop</div>
+        <div class="cart">
+            <form action="#" class="search-form">
+                <input type="text" class="search-field">
+                <button class="btn-search">
+                    <i class="fas fa-search"></i>
+                </button>
+            </form>
+            <button class="btn-cart">Cart</button>
+            <div class="cart-block ">
+                <div class="d-flex">
+                    <strong class="d-block">Всего товаров</strong>
+                    <div id="quantity"></div>
+                </div>
+                <hr>
+                <div class="cart-items">
+
+                </div>
+                <hr>
+                <div class="d-flex">
+                    <strong class="d-block">Общая ст-ть:</strong>
+                    <div id="price"></div>
+                </div>
+            </div>
+        </div>
+    </header>
+    <main>
+        <div class="products">
+        </div>
+    </main>
+
+    <script defer src="https://use.fontawesome.com/releases/v5.7.2/js/all.js" integrity="sha384-0pzryjIRos8mFBWMzSSZApWtPl/5++eIfzYmTgBBmXYdhvxPc+XcFEk+zJwDgWbP" crossorigin="anonymous"></script>
+
+</body>
+
+</html>
