+++ Students/Nina_Dokukina/Project/1/src/index.js
@@ -1,0 +1,5 @@
+import './assets/style/style.css';
+import './assets/style/normalize.css';
+
+import app from './components/main.js';
+app();
